Update web-s.py
Instead of using the pdf-parser that initially was the goal, I was assigned to write a program that can take in a .txt file and convert it to csv or otherwise properly formatted document. The particular file that I am working with had a couple of required characterizations: name, ticker, statistic, equity, bio, and then sdg categories numbered 1 - 17. Initially I was struggling a lot with writing this code because I did not know where to start with .txt parsing, luckily I was able to use the .io library in Python in order to accomplish my needs. I assigned each one of the lists to a key within a dictionary to make insertion of values more easily (because it is required to iterate over all 17 values every single time we have a new company) which made my code a lot more simpler. Biggest issue as of right now is the lack in uniformity of the representation of the data that is in the .txt file; theoretically if they were all formatted properly then the lists would be accepting/processing the data properly. One goal for next time would be to work with a smaller sample of the .txt file (because going through 50k lines is a lot), and then using that smaller sample to work on formatting (this is where Pandas comes in!).