Update web-s.py
Worked with a smaller partition of the initial .txt file in order to properly execute a function that would write to an Excel file. Initially issues w/ reaching the end of the file (would throw an IndexOutOfBound Error) which was successfully dealt with via the implementation of break statements after checking for zero length. Finally, used Pandas DataFrame to convert the lists to a structured frame and wrote to an Excel file.