Add files via upload
Worked on a new project,  which is to take data from a particular webpage and create a CSV or Excel sheet that organized the following information: state, company name, and jobs created. Started by assuming all the text which worked successfully, however running into trouble when trying to extract the specific keywords of data from the larger blocks of text due to multiple usage of state names (ex. Texas Instruments is the name of the company but Utah is the state) - issue to be resolved.