Update industrySelectScraper.py
Was able to figure out a way that can pull the correct states, of course this method is contingent on certain ways that the original text is written but it is a good start. Having a harder time trying to parse and retrieve the correct institution name from the original text because it tends to partition text unnecessarily or recognize the goal entity under the wrong tag (e.g. Organization vs Person). I do not know how I am going to get the number of jobs, this could potentially be a task to start firsthand because it will take more time.