Update industrySelectScraper.py
Managed to write program that can find the specified "state" in a particular text, however due to the plurality of identifiers for states (i.e. abbreviation vs full form), the program tends to incorrectly pick up on "South Dakota" (abbreviated as "SD") and does not parse the text for the exact keyword but rather if the keyword is represented as a substring or not. In addition, when multiple states are mentioned sometimes the program may output the wrong state, which is why instead it outputs a list of states that have the highest "count" within the original text and then appends each list to a final list, "state", which will be used to convert the text into a more readable format (such as csv or .xml).